requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# For local models
transformers
datasets

# For RAG
openai
sentence-transformers
faiss-cpu
boto3

# Utilities
pyyaml
pandas
scikit-learn
tqdm
seqeval
pytest